Formatting + Editing
The article format - while rendering correctly - was not ideal. I recommend converting whatever file you are given to a file more easily manageable, such as converting a Word file to a Google Doc. This should make converting the article into markdown much easier.

Once again, a lot of run on sentences. I also split a paragraph that was rather long.